Change from manual upload to automatic upload
Journal will now automatically upload changes made to the blog to the specified server. This means an update of the server files doesn't need to be triggered manually anymore but Journal will automatically upload files to the server if a post has been created or changed, the theme has been changed etc.